includes my La Luz pack run from today.
I'm pretty happy with my pace, considering how hot it was.  I
certainly drank all 40 oz. of my Tailwind water.  I didn't trip.
There was a very nice amount of snow in that it filled in the cracks
toward the top, although there were a few patches of gray ice.