includes my speed run from today.
I'm pretty happy with it, considering the state of my right foot.
Granted, I took 400mg of ibuprofen and 500mg of acetaminophen, but I
believe it chirped less today than a week ago.  This is not
particularly close to my best time for this particular exercise, *but*
it's the second best time for this exercise since I started logging
them in 2017.  That's primarily due to me not being injured (other
than the hot-spot) or sick (probably due to not drinking alcohol).