adds taper the week before The Quad and recovery the week after.
I did that in 2019 and I had great fitness in 2019.  With my hot-spot,
adding in a taper and recovery section will stress my foot less, but
I've also chosen to add in some pack interval training, but only 6
intervals during taper and recovery and with me parking at the library
and only doing a single loop around Albuquerque Academy.

I think giving my foot some rest is pretty important, but I don't want
to become soft.  As it is, I do not plan on pushing it hard at The
Quad and that might be a mistake.  I've never done The Quad with a
hot-spot before, so I don't know if slowing down will be a net win for
the foot or not.  My guess is it will be, however, it'll be a longer
amount of time on my feet.