includes today's La Luz run.
I was running late and although I weighed myself wearing my pack, I
forgot to weigh myself without the pack.  In fact, I went from the
scale to my truck to leave the pack there, thinking that the weight
was going to be within a pound of being correct regardless.  However,
I forgot that I do my La Luz training with 35 pounds, but all the
other trainings with 40. Oops.

So, when I started my run, it initially seemed harder than I expected
it to be.  Then it dawned on me what I had done.  Seeing how I'm
ostensibly babying my right foot, I kind of lost my willpower to push
myself hard at the lower elevation and the upper elevation had enough
snow to slow me down a bit.  So, although I was hoping to break a
fifteen minute mile average pace, I instead averaged 16:41.  Ugh!

To be fair to myself, in 2019, I did this training with a 35 pound
pack and averaged a 16:42 min/mile with an average heart rate of
131bpm. So, I'm not necessarily greatly out of shape, but this
particular training is hard to judge even when I'm using the right
amount of weight because the trail conditions vary so much from year
to year.

For the record, although I could feel my hot-spot with every step, I
don't think it was ever sufficiently uncomfortable to slow me down.