includes today's 24 mile 41# pack run.
This is the same training run I did on January 29th. In a normal training
year I'd be faster, but my hot-spot got worse and not only was it painful,
I was initially under the impression that my intervening training is what
cause it to be worse, but then I remembered...

...sometime in the last week I realized there was a fairly large bump around
my hot-spot.  I thought there was some chance that my hot-spot was a callus,
so I used a file to file it down quite a bit.  That actually seemed to make
things better.  However, at today's turnaround, I looked and saw that due
to my thinner skin, I could see a little brown dot.  That dot is either
dried blood from a puncture that I sustained at some point or is actually
foreign matter.  I'm going to find out later today which it is.

My *guess* is I have more than one problem with that foot and that
filing down the callous only made things worse.  There's also a chance
that trying to figure out what that brown dot is will make things
further worse, but I think more likely than not I'll either improve
the situation or effectively not make it better or worse (famous last
words).

It's clear to me that I'm now moving slowly not only due to the pain
in my foot, but also due to decrease fitness from the lack of running
volume and lack of speedwork during this training period.  I don't
think I'm in danger of DNF'ing (or DNS'ing) Western States, but I will
not be signing up for the Tough Ruck this year, and I consequently, I
may skip the Boston Marathon, since flying out to Boston just to do a
slow Boston Marathon on a gimpy foot may not be a good use of money.