includes today's interval pack run, updates March 25th.
My tentative plan is to see if I can get 50 miles in for the Academy
of Pain, but with about 15 of them being running to Bombs Away,
running the Burque Brewery Tour Crowning, then running back to Academy
of Pain.  I don't know if they'll count my non AA miles, but I don't
think I'll be able to crank out 50 AA miles.

I may be kidding myself and may not be up for much six days after
BMDM.  In fact, if I am up for much that may be because the wheels
fell off @ BMDM.